% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_test_prob_overall}
\alias{get_test_prob_overall}
\title{Given tree tables and model weights, obtain predicted probabilities for test data.}
\usage{
get_test_prob_overall(weights, num_cats, testdata, treetable_list)
}
\description{
Given tree tables and model weights, obtain predicted probabilities for test data.
}
