% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sBayesRF}
\alias{sBayesRF}
\title{Safe-Bayesian Random Forest. Initial test function.}
\usage{
sBayesRF(lambda, num_trees, seed, num_cats, y, original_datamat,
  alpha_parameters, beta_par, test_datamat)
}
\description{
Safe-Bayesian Random Forest. Initial test function.
}
