% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImpScores_sBF.R
\name{varImpScores_sBF}
\alias{varImpScores_sBF}
\title{Variable importances as defined by Hernandez et al. (2018)}
\usage{
varImpScores_sBF(object)
}
\arguments{
\item{object}{A sBF object.}
}
\value{
A vector of variable importances. The variables are ordered in the same order that they occur in columns of the input covariate matrix used to obtain the input BART-IS object.
}
\description{
This measure defines the importance of a variable as the model-probability weighted sum of the number of splits on the variable of interest, divided by the sum over all variables of such weighted counts of splits.
}
\examples{
#set the seed
set.seed(100)
#simulate some data

#Train the object
#Obtain the variable importances

}
