% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varIncProb_sBF.R
\name{varIncProb_sBF}
\alias{varIncProb_sBF}
\title{Variable inclusion probabilities as defined by Linero (2018)}
\usage{
varIncProb_sBF(object)
}
\arguments{
\item{object}{A sBF object obtained using the train_BART_IS or train_BART_IS_no_output function.}
}
\value{
A vector of posterior inclusion probabilities. The variables are ordered in the same order that they occur in columns of the input covariate matrix used to obtain the input sBF object.
}
\description{
This measure defines the posterior inclusion probability of a variable as the model-probability weighted sum of indicator variables for whether the variable was used in any splitting rules.
}
\examples{
#set the seed
set.seed(100)
#simulate some data
#Train the object

#Obtain the variable importances

}
